darknet-yolov4/inference/requirements.txt
flask
requests
retrying
Pillow
sagemaker-inference
multi-model-server
mxnet
boto3

#opencv-python
#numpy